Moved login, sign up, and logout to a new component called authentication.
